Final Game and GUI
Final version of the game and GUI.  Made the program slightly more
efficient.  The comments are more in depth and easier to understand and
typos removed.  The game will forever be able to be improved but this
is a solid working game that has a user friendly and aesthetically
pleasing GUI.